Created a login diner file which holds the login form for diners, and adjusted the login to display one form or another.
